v1.0.1 The Clock and the Weather has been moved to the bottom left of the launcher
Features
- Clock and weather has a Flipping effect when switching

v1.0.2
- Added the display of SSID and Password for Tethering at the bottom right corner of the Launcher overlapping the Collar Text
- Fixed the flipping 90 degree bug

v1.0.3
- Added the Display of Hotel Logo
- Created the flipping effect for Clock, Weather and Hotel Logo and also fixed the bug for flipping
- Added the pause and resume effect for the flipping effect when launcher goes in the background

v1.0.4
- Added The Unzip of TV Channels when launcher is loaded. This happens only once after the box boot up

v1.0.5
- Added Multi-Lingual Support. Launcher Menu Items, Sub Menu items and Day/Month are set to be multi Lingual
- Added World Clock
- Fixed the Screensaver Bug
- Splitted the Weather as Temperature and Remark
- Changed the Hotspot Flipping effect to Alpha effect
- Changed the Launcher Menu Design

v1.0.6d
- This Launcher is based on the 5.1 Launcher concepts
- Digital Signage now executes under AsyncTask when loading the new wallpaper, so it does not run on main thread
- Launcher has non-translucent theme which eliminates the issue of youtube running in background
- PerfectTime fix for time setting as per android 6.0
- No checking of permissions
- Default wallpaper set to black color background jpeg
- Flipping pause and resume in proper sequence and with priority so that overflipping does not happen
- Removed the dropdown delay of 250ms for subitems

v1.0.7
- Changes made for the Android 6.0
- No killing of Apps onResume()
- onResume() will not execute all its functions within 15 seconds of its calling
- The flippers will be started in the onCreate() and they wont be paused and resumed
- TV Restore will take place inside onCreate()
- Youtube restore will support older as well as newer versions of youtube
- Have support for PerfectTimeService date command for Android 6.0
- Launcher items will be processed inside the Handler()

+++ app/src/main/java/com/excel/appstvlauncher/secondgen/MainActivity.java
@@ -41,6 +41,7 @@ import com.excel.flipper.Flipper;
 import com.excel.imagemanipulator.DigitalSignage;
 import com.excel.imagemanipulator.DigitalSignageHolder;
 import com.excel.perfecttime.PerfectTimeService;
+import com.excel.util.MD5;
 import com.excel.yahooweather.Weather;
 
 import org.json.JSONArray;
@@ -657,7 +658,7 @@ public class MainActivity extends Activity {
 
 					startScreenCastService();
 
-					UtilShell.executeShellCommand( "am force-stop com.google.android.youtube.tv" );
+					//UtilShell.executeShellCommand( "am force-stop com.google.android.youtube.tv" );
 				}
 			}
 		}
@@ -1395,24 +1396,70 @@ public class MainActivity extends Activity {
     public void unzipTvChannelsZip(){
         Log.i( TAG, "unzipTvChannelsZip() executed" );
 
-        UtilShell.executeShellCommand( "rm -r /mnt/sdcard/appstv_data/tv_channels/backup",
-                "unzip -o /mnt/sdcard/appstv_data/tv_channels/tv_channels.zip -d /mnt/sdcard/appstv_data/tv_channels" );
+		UtilShell.executeShellCommandWithOp( "rm -r /mnt/sdcard/appstv_data/tv_channels/backup",
+				"unzip -o /mnt/sdcard/appstv_data/tv_channels/tv_channels.zip -d /mnt/sdcard/appstv_data/tv_channels" );
 
-        // 1. kill com.android.dtv
-        String pid = UtilShell.executeShellCommandWithOp( "pidof com.android.dtv" );
-        //UtilShell.executeShellCommandWithOp( "kill "+pid );
+		// Compare the file program.db
+		File programdb_internal = new File( "/data/hdtv/program.db" );
+		File programdb_sdcard	= new File( "/mnt/sdcard/appstv_data/tv_channels/backup/hdtv/program.db" );
 
-        UtilShell.executeShellCommand( "chmod -R 777 /data/hdtv",
-                "rm -r /data/hdtv/*",
-                "cp -r /mnt/sdcard/appstv_data/tv_channels/backup/hdtv/* /data/hdtv",
-                "chmod -R 777 /data/hdtv" );
 
-        // last. kill com.android.dtv
-        pid = UtilShell.executeShellCommandWithOp( "pidof com.android.dtv" );
-        //UtilShell.executeShellCommandWithOp( "kill "+pid );
+		if( !programdb_sdcard.exists() ) {
+			Log.e( TAG, "Restore terminated, program.db file does not exist on sdcard" );
+			setTvChannelRestored( true );
+			return;
+		}
 
-        setTvChannelRestored( true );
-        Log.d( TAG, "tv_channels.zip extracted successfully" );
+		try {
+
+			String md5_internal = "";
+			String md5_sdcard 	= MD5.getMD5Checksum( programdb_sdcard );
+
+			if( !programdb_internal.exists() ) {
+				md5_internal = "xxxxx";
+				Log.d( TAG, "Internal program.db does not exist" );
+				//return;
+			}
+			else{
+				md5_internal = MD5.getMD5Checksum( programdb_internal );
+			}
+
+			Log.d( TAG, md5_internal + "," +md5_sdcard );
+
+			// If the two files do not match, then restore
+			if( ! md5_internal.equals( md5_sdcard ) ) {
+
+
+				// 1. kill com.android.dtv
+				//String pid = UtilShell.executeShellCommandWithOp( "pidof com.android.dtv" );
+				//UtilShell.executeShellCommandWithOp( "kill "+pid );
+
+				UtilShell.executeShellCommand( "chmod -R 777 /data/hdtv",
+						"rm -r /data/hdtv/*",
+						"cp -r /mnt/sdcard/appstv_data/tv_channels/backup/hdtv/* /data/hdtv",
+						"chmod -R 777 /data/hdtv" );
+
+				Log.d( TAG, "tv_channels.zip extracted successfully" );
+
+				// last. kill com.android.dtv
+				//pid = UtilShell.executeShellCommandWithOp("pidof com.android.dtv");
+				//UtilShell.executeShellCommandWithOp( "kill "+pid );
+
+				Log.i( TAG, "Restored tv channels successfully !" );
+
+				//setTvChannelRestored( true );
+			}
+			else{
+				Log.i( TAG, "TV channels not restored because they are same !" );
+
+			}
+			setTvChannelRestored( true );
+		}
+		catch ( Exception e ){
+			e.printStackTrace();
+		}
+
+
     }
 
     public void restoreTvChannels(){
